api/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
grpcio==1.74.0
grpcio-tools==1.74.0
minio==7.2.16
requests==2.32.5
psycopg2-binary==2.9.10
Flask-Pydantic==0.13.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
python-dotenv